% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plota_rede.R
\name{explora}
\alias{explora}
\title{Exploration plot}
\usage{
explora(
  file,
  filtro = c("Dá ordens para", "Trabalha com", "Fornece drogas para", "Intimo de",
    "protege")
)
}
\arguments{
\item{file}{Path to csv file of edges. Must be without header.}

\item{filtro}{String. Relation to filter the multiplex network. One of predefined options.}
}
\value{
a visNetwork object
}
\description{
Exploration plot
}
